fix init command.
